requirements.txt
flask==2.3.3
flask-cors==4.0.0
pdfplumber==0.10.2
Pillow==10.0.0
pytesseract==0.3.10
numpy==1.26.0
pandas==2.1.0
torch==2.0.1
transformers==4.35.0
datasets==2.14.5
scikit-learn==1.3.0